fix(vscode-insiders): add timestamp to version
VSCode Insider's "productVersion" doesn't change regularly since it tracks the
next version of VSCode, but its "timestamp" corresponds to the time that the
latest release was created, and thus changes changes nightly. This allows
Scoop's checkver mechanism to work correctly since timestamps are comparable.

Include a workaround for a bug in Newtonsoft.Json's [FieldMultipleFilter](
https://github.com/JamesNK/Newtonsoft.Json/blob/13.0.1/Src/Newtonsoft.Json/Linq/JsonPath/FieldMultipleFilter.cs),
where it always throws an exception when passing `true` as `errorWhenNoMatch`
of `JToken.SelectTokens(path, errorWhenNoMatch)` when using the JsonPath
multiple property union syntax (e.g. `$['Prop1','Prop2']`). This workaround
can be removed once the bug is resolved.